% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{options_csv_to_list}
\alias{options_csv_to_list}
\title{options_csv_to_list}
\usage{
options_csv_to_list(csv_file, to_ivol_format = T, add_stockprice = T)
}
\arguments{
\item{to_ivol_format}{}
}
\description{
Convert the CSV obtained from WRDS to same structure as dataset obtained from
ivolatility. Addditionally, add adjusted close price.
}
