% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{request}
\usage{
request(
  endpoint,
  symbol = NA_character_,
  date = NA_character_,
  from = NA_character_,
  to = NA_character_,
  region = NA_character_,
  callPut = NA_character_
)
}
\arguments{
\item{endpoint}{Endpoint (character)}

\item{symbol}{Ticker of equity or option (character)}

\item{date}{Date in format "YYYY-MM-DD" (character)}

\item{from}{Date in format "YYYY-MM-DD" (character)}

\item{to}{Date in format "YYYY-MM-DD" (character)}

\item{region}{Equity region. Options: "USA", "EUROPE", "ASIA", "CANADA", "RUSSIA", "CRYPTO" (character)}

\item{callPut}{"C" for Call or "P" for Put (character)}
}
\value{
data.frame of API response
}
\description{
Request given endpoint using parameters provided by funtion.
List of endpoints available here \url{https://redocly.github.io/redoc/?nocors&url=https://restapi.ivolatility.com/api-docs#section/Endpoints-Access}.
}
