% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling.R
\name{enrich_optionchains}
\alias{enrich_optionchains}
\title{Title}
\usage{
enrich_optionchains(df)
}
\description{
Convert a option chain data.frame into format suitable for ipod estimation
as well as enrich option chain with information relevant for estimation.
}
\details{
Finished data.frame should have following structure:
\itemize{
\item date
\item time to maturity
\item price (bid+ask/2, stock price for strike price=0)
\item risk free rate
\item open interest
\item weight (open interest_i / sum(open interest))
}
}
